requirements.txt
asgiref==3.8.1
Django==4.2.9
django-crispy-forms==2.3
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
